Dreame integration use base class
