+++ miio/__init__.py
@@ -41,8 +41,11 @@ from miio.heater import Heater
 from miio.heater_miot import HeaterMiot
 from miio.huizuo import Huizuo, HuizuoLampFan, HuizuoLampHeater, HuizuoLampScene
 from miio.integrations.petwaterdispenser import PetWaterDispenser
-from miio.integrations.vacuum.dreame.dreamevacuum_f9_miot import DreameF9VacuumMiot
-from miio.integrations.vacuum.dreame.dreamevacuum_miot import DreameVacuumMiot
+from miio.integrations.vacuum.dreame.dreamevacuum_miot import (
+    DreameC1Vacuum,
+    DreameF9Vacuum,
+    DreameVacuumMiot,
+)
 from miio.integrations.vacuum.mijia import G1Vacuum
 from miio.integrations.vacuum.roborock import RoborockVacuum, Vacuum, VacuumException
 from miio.integrations.vacuum.roborock.vacuumcontainers import (
