+++ miio/discovery.py
@@ -32,7 +32,8 @@ from . import (
     ChuangmiPlug,
     Cooker,
     Device,
-    DreameF9VacuumMiot,
+    DreameC1Vacuum,
+    DreameF9Vacuum,
     Fan,
     FanLeshow,
     FanMiot,
@@ -209,7 +210,8 @@ DEVICE_MAP: Dict[str, Union[Type[Device], partial]] = {
     "viomi-vacuum-v8": ViomiVacuum,
     "zhimi.heater.za1": partial(Heater, model=MODEL_HEATER_ZA1),
     "zhimi.elecheater.ma1": partial(Heater, model=MODEL_HEATER_MA1),
-    "dreame-vacuum-p2008": DreameF9VacuumMiot,
+    "dreame-vacuum-mc1808": DreameC1Vacuum,
+    "dreame-vacuum-p2008": DreameF9Vacuum,
 }
