+++ miio/integrations/vacuum/dreame/dreamevacuum_miot.py
@@ -1,40 +1,97 @@
-"""Vacuum 1C STYTJ01ZHM (dreame.vacuum.mc1808)"""
+"""Dreame Vacuum."""
 
 import logging
 from enum import Enum
+from typing import Dict
+
+import click
 
 from miio.click_common import command, format_output
+from miio.exceptions import DeviceException
 from miio.miot_device import DeviceStatus as DeviceStatusContainer
 from miio.miot_device import MiotDevice, MiotMapping
+from miio.utils import deprecated
 
 _LOGGER = logging.getLogger(__name__)
 
-_MAPPING: MiotMapping = {
-    "battery_level": {"siid": 2, "piid": 1},
-    "charging_state": {"siid": 2, "piid": 2},
-    "device_fault": {"siid": 3, "piid": 1},
-    "device_status": {"siid": 3, "piid": 2},
-    "brush_left_time": {"siid": 26, "piid": 1},
-    "brush_life_level": {"siid": 26, "piid": 2},
-    "filter_life_level": {"siid": 27, "piid": 1},
-    "filter_left_time": {"siid": 27, "piid": 2},
-    "brush_left_time2": {"siid": 28, "piid": 1},
-    "brush_life_level2": {"siid": 28, "piid": 2},
-    "operating_mode": {"siid": 18, "piid": 1},
-    "cleaning_mode": {"siid": 18, "piid": 6},
-    "delete_timer": {"siid": 18, "piid": 8},
-    "life_sieve": {"siid": 19, "piid": 1},
-    "life_brush_side": {"siid": 19, "piid": 2},
-    "life_brush_main": {"siid": 19, "piid": 3},
-    "timer_enable": {"siid": 20, "piid": 1},
-    "start_time": {"siid": 20, "piid": 2},
-    "stop_time": {"siid": 20, "piid": 3},
-    "deg": {"siid": 21, "piid": 1, "access": ["write"]},
-    "speed": {"siid": 21, "piid": 2, "access": ["write"]},
-    "map_view": {"siid": 23, "piid": 1},
-    "frame_info": {"siid": 23, "piid": 2},
-    "volume": {"siid": 24, "piid": 1},
-    "voice_package": {"siid": 24, "piid": 3},
+
+DREAME_C1 = "dreame.vacuum.mc1808"
+DREAME_F9 = "dreame.vacuum.p2008"
+
+MIOT_MAPPING: Dict[str, MiotMapping] = {
+    DREAME_C1: {
+        # https://home.miot-spec.com/spec/dreame.vacuum.mc1808
+        "battery_level": {"siid": 2, "piid": 1},
+        "charging_state": {"siid": 2, "piid": 2},
+        "device_fault": {"siid": 3, "piid": 1},
+        "device_status": {"siid": 3, "piid": 2},
+        "brush_left_time": {"siid": 26, "piid": 1},
+        "brush_life_level": {"siid": 26, "piid": 2},
+        "filter_life_level": {"siid": 27, "piid": 1},
+        "filter_left_time": {"siid": 27, "piid": 2},
+        "brush_left_time2": {"siid": 28, "piid": 1},
+        "brush_life_level2": {"siid": 28, "piid": 2},
+        "operating_mode": {"siid": 18, "piid": 1},
+        "cleaning_mode": {"siid": 18, "piid": 6},
+        "delete_timer": {"siid": 18, "piid": 8},
+        "life_sieve": {"siid": 19, "piid": 1},
+        "life_brush_side": {"siid": 19, "piid": 2},
+        "life_brush_main": {"siid": 19, "piid": 3},
+        "timer_enable": {"siid": 20, "piid": 1},
+        "start_time": {"siid": 20, "piid": 2},
+        "stop_time": {"siid": 20, "piid": 3},
+        "deg": {"siid": 21, "piid": 1, "access": ["write"]},
+        "speed": {"siid": 21, "piid": 2, "access": ["write"]},
+        "map_view": {"siid": 23, "piid": 1},
+        "frame_info": {"siid": 23, "piid": 2},
+        "volume": {"siid": 24, "piid": 1},
+        "voice_package": {"siid": 24, "piid": 3},
+        "timezone": {"siid": 25, "piid": 1},
+        "home": {"siid": 2, "aiid": 1},
+        "locate": {"siid": 17, "aiid": 1},
+        "start_clean": {"siid": 3, "aiid": 1},
+        "stop_clean": {"siid": 3, "aiid": 2},
+        "reset_mainbrush_life": {"siid": 26, "aiid": 1},
+        "reset_filter_life": {"siid": 27, "aiid": 1},
+        "reset_sidebrush_life": {"siid": 28, "aiid": 1},
+        "move": {"siid": 21, "aiid": 1},
+        "play_sound": {"siid": 24, "aiid": 3},
+    },
+    DREAME_F9: {
+        # https://home.miot-spec.com/spec/dreame.vacuum.p2008
+        "battery_level": {"siid": 3, "piid": 1},
+        "charging_state": {"siid": 3, "piid": 2},
+        "device_fault": {"siid": 2, "piid": 2},
+        "device_status": {"siid": 2, "piid": 1},
+        "brush_left_time": {"siid": 9, "piid": 1},
+        "brush_life_level": {"siid": 9, "piid": 2},
+        "filter_life_level": {"siid": 11, "piid": 1},
+        "filter_left_time": {"siid": 11, "piid": 2},
+        "brush_left_time2": {"siid": 10, "piid": 1},
+        "brush_life_level2": {"siid": 10, "piid": 2},
+        "operating_mode": {"siid": 4, "piid": 1},
+        "cleaning_mode": {"siid": 4, "piid": 4},
+        "delete_timer": {"siid": 18, "piid": 8},
+        "timer_enable": {"siid": 5, "piid": 1},
+        "start_time": {"siid": 5, "piid": 2},
+        "stop_time": {"siid": 5, "piid": 3},
+        "map_view": {"siid": 6, "piid": 1},
+        "frame_info": {"siid": 6, "piid": 2},
+        "volume": {"siid": 7, "piid": 1},
+        "voice_package": {"siid": 7, "piid": 2},
+        "water_flow": {"siid": 4, "piid": 5},
+        "water_tank_status": {"siid": 4, "piid": 6},
+        "timezone": {"siid": 8, "piid": 1},
+        "home": {"siid": 3, "aiid": 1},
+        "locate": {"siid": 7, "aiid": 1},
+        "start_clean": {"siid": 4, "aiid": 1},
+        "stop_clean": {"siid": 4, "aiid": 2},
+        "reset_mainbrush_life": {"siid": 9, "aiid": 1},
+        "reset_filter_life": {"siid": 11, "aiid": 1},
+        "reset_sidebrush_life": {"siid": 10, "aiid": 1},
+        "move": {"siid": 21, "aiid": 1},
+        "play_sound": {"siid": 7, "aiid": 2},
+    },
 }
 
 
@@ -46,12 +103,20 @@ class ChargingState(Enum):
     GoCharging = 5
 
 
-class CleaningMode(Enum):
+class CleaningModeDreameC1(Enum):
     Unknown = -1
     Quiet = 0
     Default = 1
     Medium = 2
     Strong = 3
+
+
+class CleaningModeDreameF9(Enum):
+    Unknown = -1
+    Quiet = 0
+    Standart = 1
+    Strong = 2
+    Turbo = 3
 
 
 class OperatingMode(Enum):
@@ -82,7 +147,20 @@ class DeviceStatus(Enum):
     ManualSweeping = 13
 
 
-class DreameVacuumStatus(DeviceStatusContainer):
+class WaterFlow(Enum):
+    Unknown = -1
+    Low = 1
+    Medium = 2
+    High = 3
+
+
+class WaterTankStatus(Enum):
+    Unknown = -1
+    NotAttached = 0
+    Attached = 1
+
+
+class DreameVacuumStatusBase(DeviceStatusContainer):
     def __init__(self, data):
         self.data = data
 
@@ -139,32 +217,12 @@ class DreameVacuumStatus(DeviceStatusContainer):
             return OperatingMode.Unknown
 
     @property
-    def cleaning_mode(self) -> CleaningMode:
-        try:
-            return CleaningMode(self.data["cleaning_mode"])
-        except ValueError:
-            _LOGGER.error("Unknown CleaningMode (%s)", self.data["cleaning_mode"])
-            return CleaningMode.Unknown
-
-    @property
     def device_status(self) -> DeviceStatus:
         try:
             return DeviceStatus(self.data["device_status"])
         except TypeError:
             _LOGGER.error("Unknown DeviceStatus (%s)", self.data["device_status"])
             return DeviceStatus.Unknown
-
-    @property
-    def life_sieve(self) -> str:
-        return self.data["life_sieve"]
-
-    @property
-    def life_brush_side(self) -> str:
-        return self.data["life_brush_side"]
-
-    @property
-    def life_brush_main(self) -> str:
-        return self.data["life_brush_main"]
 
     @property
     def timer_enable(self) -> str:
@@ -190,11 +248,165 @@ class DreameVacuumStatus(DeviceStatusContainer):
     def voice_package(self) -> str:
         return self.data["voice_package"]
 
+    @property
+    def timezone(self) -> str:
+        return self.data["timezone"]
 
-class DreameVacuumMiot(MiotDevice):
+
+class DreameC1VacuumStatus(DreameVacuumStatusBase):
+    @property
+    def cleaning_mode(self) -> CleaningModeDreameC1:
+        try:
+            return CleaningModeDreameC1(self.data["cleaning_mode"])
+        except ValueError:
+            _LOGGER.error("Unknown CleaningMode (%s)", self.data["cleaning_mode"])
+            return CleaningModeDreameC1.Unknown
+
+    @property
+    def life_sieve(self) -> str:
+        return self.data["life_sieve"]
+
+    @property
+    def life_brush_side(self) -> str:
+        return self.data["life_brush_side"]
+
+    @property
+    def life_brush_main(self) -> str:
+        return self.data["life_brush_main"]
+
+
+class DreameF9VacuumStatus(DreameVacuumStatusBase):
+    @property
+    def water_flow(self) -> WaterFlow:
+        try:
+            return WaterFlow(self.data["water_flow"])
+        except ValueError:
+            _LOGGER.error("Unknown WaterFlow (%s)", self.data["water_flow"])
+            return WaterFlow.Unknown
+
+    @property
+    def water_tank_status(self) -> WaterTankStatus:
+        try:
+            return WaterTankStatus(self.data["water_tank_status"])
+        except ValueError:
+            _LOGGER.error("Unknown WaterFlow (%s)", self.data["water_tank_status"])
+            return WaterTankStatus.Unknown
+
+    @property
+    def cleaning_mode(self) -> CleaningModeDreameF9:
+        try:
+            return CleaningModeDreameF9(self.data["cleaning_mode"])
+        except ValueError:
+            _LOGGER.error("Unknown CleaningMode (%s)", self.data["cleaning_mode"])
+            return CleaningModeDreameF9.Unknown
+
+
+class DreameVacuumBase(MiotDevice):
+    # TODO: check the actual limit for this
+    MANUAL_ROTATION_MAX = 120
+    MANUAL_ROTATION_MIN = -MANUAL_ROTATION_MAX
+    MANUAL_DISTANCE_MAX = 300
+    MANUAL_DISTANCE_MIN = -300
+
+    @command()
+    def start(self) -> None:
+        """Start cleaning."""
+        return self.call_action("start_clean")
+
+    @command()
+    def stop(self) -> None:
+        """Stop cleaning."""
+        return self.call_action("stop_clean")
+
+    @command()
+    def home(self) -> None:
+        """Return to home."""
+        return self.call_action("home")
+
+    @command()
+    def identify(self) -> None:
+        """Locate the device (i am here)."""
+        return self.call_action("locate")
+
+    @command()
+    def reset_mainbrush_life(self) -> None:
+        """Reset main brush life."""
+        return self.call_action("reset_mainbrush_life")
+
+    @command()
+    def reset_filter_life(self) -> None:
+        """Reset filter life."""
+        return self.call_action("reset_filter_life")
+
+    @command()
+    def reset_sidebrush_life(self) -> None:
+        """Reset side brush life."""
+        return self.call_action("reset_sidebrush_life")
+
+    @command()
+    def play_sound(self) -> None:
+        """Play sound."""
+        return self.call_action("play_sound")
+
+    @command(
+        click.argument("distance", default=30, type=int),
+    )
+    def forward(self, distance: int) -> None:
+        """Move forward."""
+        if distance < self.MANUAL_DISTANCE_MIN or distance > self.MANUAL_DISTANCE_MAX:
+            raise DeviceException(
+                "Given distance is invalid, should be [%s, %s], was: %s"
+                % (self.MANUAL_DISTANCE_MIN, self.MANUAL_DISTANCE_MAX, distance)
+            )
+        self.call_action(
+            "move",
+            [
+                {
+                    "piid": 1,
+                    "value": "0",
+                },
+                {
+                    "piid": 2,
+                    "value": f"{distance}",
+                },
+            ],
+        )
+
+    @command(
+        click.argument("rotatation", default=90, type=int),
+    )
+    def rotate(self, rotatation: int) -> None:
+        """Rotate vacuum."""
+        if (
+            rotatation < self.MANUAL_ROTATION_MIN
+            or rotatation > self.MANUAL_ROTATION_MAX
+        ):
+            raise DeviceException(
+                "Given rotation is invalid, should be [%s, %s], was %s"
+                % (self.MANUAL_ROTATION_MIN, self.MANUAL_ROTATION_MAX, rotatation)
+            )
+        self.call_action(
+            "move",
+            [
+                {
+                    "piid": 1,
+                    "value": f"{rotatation}",
+                },
+                {
+                    "piid": 2,
+                    "value": "0",
+                },
+            ],
+        )
+
+
+class DreameC1Vacuum(DreameVacuumBase):
     """Interface for Vacuum 1C STYTJ01ZHM (dreame.vacuum.mc1808)"""
 
-    mapping = _MAPPING
+    _supported_models = [
+        DREAME_C1,
+    ]
+    mapping = MIOT_MAPPING[DREAME_C1]
 
     @command(
         default_output=format_output(
@@ -215,6 +427,7 @@ class DreameVacuumMiot(MiotDevice):
             "Operating mode: {result.operating_mode.name}\n"
             "Side cleaning brush left time: {result.brush_left_time2}\n"
             "Side cleaning brush life level: {result.brush_life_level2}\n"
+            "Time zone: {result.timezone}\n"
             "Timer enabled: {result.timer_enable}\n"
             "Timer start time: {result.start_time}\n"
             "Timer stop time: {result.stop_time}\n"
@@ -222,61 +435,65 @@ class DreameVacuumMiot(MiotDevice):
             "Volume: {result.volume}\n",
         )
     )
-    def status(self) -> DreameVacuumStatus:
+    def status(self) -> DreameC1VacuumStatus:
         """State of the vacuum."""
 
-        return DreameVacuumStatus(
+        return DreameC1VacuumStatus(
             {
                 prop["did"]: prop["value"] if prop["code"] == 0 else None
                 for prop in self.get_properties_for_mapping(max_properties=10)
             }
         )
 
-    def send_action(self, siid, aiid, params=None):
-        """Send action to device."""
 
-        # {"did":"<mydeviceID>","siid":18,"aiid":1,"in":[{"piid":1,"value":2}]
-        if params is None:
-            params = []
-        payload = {
-            "did": f"call-{siid}-{aiid}",
-            "siid": siid,
-            "aiid": aiid,
-            "in": params,
-        }
-        return self.send("action", payload)
+class DreameVacuumMiot(DreameC1Vacuum):
+    @deprecated(
+        "This class is replaced with DreameC1Vacuum. Use DreameC1Vacuum to control Dreame C1 vacuums."
+    )
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 
-    @command()
-    def start(self) -> None:
-        """Start cleaning."""
-        return self.send_action(3, 1)
 
-    @command()
-    def stop(self) -> None:
-        """Stop cleaning."""
-        return self.send_action(3, 2)
+class DreameF9Vacuum(DreameVacuumBase):
+    """Interface for Vacuum F9 (dreame.vacuum.p2008)"""
 
-    @command()
-    def home(self) -> None:
-        """Return to home."""
-        return self.send_action(2, 1)
+    _supported_models = [
+        DREAME_F9,
+    ]
+    mapping = MIOT_MAPPING[DREAME_F9]
 
-    @command()
-    def identify(self) -> None:
-        """Locate the device (i am here)."""
-        return self.send_action(17, 1)
+    @command(
+        default_output=format_output(
+            "\n",
+            "Battery level: {result.battery_level}\n"
+            "Brush life level: {result.brush_life_level}\n"
+            "Brush left time: {result.brush_left_time}\n"
+            "Charging state: {result.charging_state.name}\n"
+            "Cleaning mode: {result.cleaning_mode.name}\n"
+            "Device fault: {result.device_fault.name}\n"
+            "Device status: {result.device_status.name}\n"
+            "Filter left level: {result.filter_left_time}\n"
+            "Filter life level: {result.filter_life_level}\n"
+            "Map view: {result.map_view}\n"
+            "Operating mode: {result.operating_mode.name}\n"
+            "Side cleaning brush left time: {result.brush_left_time2}\n"
+            "Side cleaning brush life level: {result.brush_life_level2}\n"
+            "Time zone: {result.timezone}\n"
+            "Timer enabled: {result.timer_enable}\n"
+            "Timer start time: {result.start_time}\n"
+            "Timer stop time: {result.stop_time}\n"
+            "Voice package: {result.voice_package}\n"
+            "Volume: {result.volume}\n"
+            "Water flow: {result.water_flow.name}\n"
+            "Water tank status: {result.water_tank_status.name}\n",
+        )
+    )
+    def status(self) -> DreameF9VacuumStatus:
+        """State of the vacuum."""
 
-    @command()
-    def reset_mainbrush_life(self) -> None:
-        """Reset main brush life."""
-        return self.send_action(26, 1)
-
-    @command()
-    def reset_filter_life(self) -> None:
-        """Reset filter life."""
-        return self.send_action(27, 1)
-
-    @command()
-    def reset_sidebrush_life(self) -> None:
-        """Reset side brush life."""
-        return self.send_action(28, 1)
+        return DreameF9VacuumStatus(
+            {
+                prop["did"]: prop["value"] if prop["code"] == 0 else None
+                for prop in self.get_properties_for_mapping(max_properties=10)
+            }
+        )
